requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
httpx
loguru
numpy
pandas
scikit-learn
torch
transformers
sentence-transformers
faiss-cpu
pyyaml
orjson
protobuf
typing_extensions
tqdm
mlflow
google-generativeai
psutil
gradio
